servo: Merge #3772 - layout: Shrink fragments down from 448 bytes to 128 bytes (from pcwalton:slim-down-fragment); r=metajack
16% performance improvement in layout (!)

r? metajack (or whoever)

Source-Repo: https://github.com/servo/servo
Source-Revision: 012a80cc182ea9773f82696bf36f05820ab994b3

UltraBlame original commit: a121717fa84101bbc9b2a9ee2216b36d43d94f9b

+++ servo/components/layout/layout_debug.rs
@@ -92,8 +92,8 @@ impl Drop for Scope {
 
 
 
-pub fn generate_unique_debug_id() -> uint {
-    unsafe { DEBUG_ID_COUNTER.fetch_add(1, SeqCst) }
+pub fn generate_unique_debug_id() -> u16 {
+    unsafe { DEBUG_ID_COUNTER.fetch_add(1, SeqCst) as u16 }
 }
